resolve URLs passed to hyperquest
The normalized this.writeNpm.href has an ending slash. As a slash was
appended to this string (along with the package name), the URL passed to
hyperquest contained two slashes after the hostname. For example:

    http://registry.npmjs.com//smart-private-npm

The router used by registry3 now responds with a 500 error to this type
of request.

By resolving the package name with the registry URL, we obtain URLs that
the registry3 router accepts.

    http://registry.npmjs.com/smart-private-npm